Fixed ~300 errors in rebasing, giving some TLC to KAGIC gems. DO NOT BUILD THIS IT WILL BREAK!

+++ src/main/java/mod/amalgam/client/model/fusions/ModelMalachite.java
@@ -202,7 +202,7 @@ public class ModelMalachite extends ModelBiped {
 			f1 = 1.0F - f1;
 			float f2 = MathHelper.sin(f1 * (float)Math.PI);
 			float f3 = MathHelper.sin(this.swingProgress * (float)Math.PI) * -(this.bipedHead.rotateAngleX - 0.7F) * 0.75F;
-			mainHandModel.rotateAngleX = (float)((double)mainHandModel.rotateAngleX - ((double)f2 * 1.2D + (double)f3));
+			mainHandModel.rotateAngleX = (float)(mainHandModel.rotateAngleX - (f2 * 1.2D + f3));
 			mainHandModel.rotateAngleY += this.bipedBody.rotateAngleY * 2.0F;
 			mainHandModel.rotateAngleZ += MathHelper.sin(this.swingProgress * (float)Math.PI) * -0.4F;
 		}
