+++ src/main/java/mod/amalgam/client/model/fusions/ModelRainbowQuartz.java
@@ -201,7 +201,7 @@ public class ModelRainbowQuartz extends ModelBiped {
 			f1 = 1.0F - f1;
 			float f2 = MathHelper.sin(f1 * (float)Math.PI);
 			float f3 = MathHelper.sin(this.swingProgress * (float)Math.PI) * -(this.bipedHead.rotateAngleX - 0.7F) * 0.75F;
-			modelrenderer.rotateAngleX = (float)((double)modelrenderer.rotateAngleX - ((double)f2 * 1.2D + (double)f3));
+			modelrenderer.rotateAngleX = (float)(modelrenderer.rotateAngleX - (f2 * 1.2D + f3));
 			modelrenderer.rotateAngleY += this.bipedBody.rotateAngleY * 2.0F;
 			modelrenderer.rotateAngleZ += MathHelper.sin(this.swingProgress * (float)Math.PI) * -0.4F;
 		}
