+++ src/main/java/mod/amalgam/client/render/layers/LayerGemPlacement.java
@@ -43,9 +43,9 @@ public class LayerGemPlacement implements LayerRenderer<EntityGem> {
 			}
 			this.gemRenderer.bindTexture(this.getTexture(gem));
 			int color = gem.getGemColor();
-			float r = (float) ((color & 16711680) >> 16) / 255f;
-	        float g = (float) ((color & 65280) >> 8) / 255f;
-	        float b = (float) ((color & 255) >> 0) / 255f;
+			float r = ((color & 16711680) >> 16) / 255f;
+	        float g = ((color & 65280) >> 8) / 255f;
+	        float b = ((color & 255) >> 0) / 255f;
 			GlStateManager.color(r, g, b);
 			this.gemModel.setModelAttributes(this.gemRenderer.getMainModel());
 	        this.gemModel.render(gem, limbSwing, limbSwingAmount, ageInTicks, netHeadYaw, headPitch, scale);
