+++ src/main/java/mod/amalgam/client/render/layers/LayerRutileGemPlacement.java
@@ -25,9 +25,9 @@ public class LayerRutileGemPlacement implements LayerRenderer<EntityRutile> {
 		if (!gem.isFusion()) {
 			this.gemRenderer.bindTexture(this.getTexture(gem));
 			int color = gem.getGemColor();
-			float r = (float) ((color & 16711680) >> 16) / 255f;
-	        float g = (float) ((color & 65280) >> 8) / 255f;
-	        float b = (float) ((color & 255) >> 0) / 255f;
+			float r = ((color & 16711680) >> 16) / 255f;
+	        float g = ((color & 65280) >> 8) / 255f;
+	        float b = ((color & 255) >> 0) / 255f;
 			GlStateManager.color(r, g, b);
 			this.gemRenderer.getModel(gem.isDefective()).setModelAttributes(this.gemRenderer.getMainModel());
 			this.gemRenderer.getModel(gem.isDefective()).render(gem, limbSwing, limbSwingAmount, ageInTicks, netHeadYaw, headPitch, scale);
