+++ src/main/java/mod/amalgam/client/render/layers/LayerSapphireItem.java
@@ -43,7 +43,7 @@ public class LayerSapphireItem implements LayerRenderer<EntitySapphire> {
 			GlStateManager.rotate(-90.0F, 1.0F, 0.0F, 0.0F);
 			GlStateManager.rotate(180.0F, 0.0F, 1.0F, 0.0F);
 			boolean flag = handSide == EnumHandSide.LEFT;
-			GlStateManager.translate((float)(flag ? -1 : 1) / 16.0F, 0.125F, -0.6F);
+			GlStateManager.translate((flag ? -1 : 1) / 16.0F, 0.125F, -0.6F);
 			Minecraft.getMinecraft().getItemRenderer().renderItemSide(entity, stack, camera, flag);
 			GlStateManager.popMatrix();
 		}
